+++ src/main/preview/index.ts
@@ -86,6 +86,7 @@ const VIDEO_EXTS = new Set([
   'divx',
   'mov',
   'wmv',
+  'asf',
   'm4v',
   'mpg',
   'mpeg',