+++ src/shared/vidThumbCache.ts
@@ -19,6 +19,7 @@ const VIDEO_EXTS = new Set([
   'divx',
   'mov',
   'wmv',
+  'asf',
   'm4v',
   'mpg',
   'mpeg',